# repos/server/static/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Storeways Open Source API",
    "version": "1.0.0",
    "description": "API documentation for storeways"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local dev server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/v1/stores/orders": {
      "get": {
        "summary": "Fetch my orders",
        "tags": [
          "Orders"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "additionalProperties": false
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "🔒 Requires JWT token. Accessible only by users with the \"customer\" role."
      }
    },
...
    "/v1/stores/payments/orders": {
      "post": {
        "summary": "Create an order",
        "tags": [
          "Payments"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number",
                    "format": "float",
                    "minimum": 1
                  },
                  "name": {
                    "type": "string"
                  },
                  "mobile": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "landmark": {
                    "type": "string"
                  },
                  "pincode": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "otp": {
                    "type": "string",
                    "minLength": 6,
                    "maxLength": 6
                  },
                  "products": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {},
                      "additionalProperties": false
                    }
                  },
                  "paymentMode": {
                    "type": "string",
                    "enum": [
                      "online",
                      "cod"
                    ]
                  }
                },
                "required": [
                  "amount",
                  "name",
                  "mobile",
                  "address",
                  "landmark",
                  "pincode",
                  "email",
                  "otp",
                  "products",
                  "paymentMode"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
  }
}